+++ mod/runtimeapi/runtime-web/src/main/java/org/onap/dcae/runtime/web/service/BlueprintInventory.java
@@ -17,6 +17,7 @@
  */
 package org.onap.dcae.runtime.web.service;
 
+import org.json.JSONException;
 import org.onap.dcae.runtime.core.FlowGraphParser.BlueprintVessel;
 import org.onap.dcae.runtime.web.models.DashboardConfig;
 import org.json.JSONObject;
@@ -74,12 +75,16 @@ public class BlueprintInventory {
 
     private JSONObject prepareBlueprintJsonObject(String blueprintName, int version, String blueprintContent) {
         JSONObject blueprintJsonObject = new JSONObject();
-        blueprintJsonObject.put("owner","dcae_mod");
-        blueprintJsonObject.put("typeName",blueprintName);
-        blueprintJsonObject.put("typeVersion",version);
-        blueprintJsonObject.put("blueprintTemplate",blueprintContent);
-        blueprintJsonObject.put("application","DCAE");
-        blueprintJsonObject.put("component","dcae");
+        try {
+            blueprintJsonObject.put("owner","dcae_mod");
+            blueprintJsonObject.put("typeName",blueprintName);
+            blueprintJsonObject.put("typeVersion",version);
+            blueprintJsonObject.put("blueprintTemplate",blueprintContent);
+            blueprintJsonObject.put("application","DCAE");
+            blueprintJsonObject.put("component","dcae");
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
         return blueprintJsonObject;
     }